fix: truncated dropdown should not determine items based on CMS position
