+++ src/open_inwoner/scss/components/Header/PrimaryNavigation.scss
@@ -154,11 +154,27 @@
     .primary-navigation__list .primary-navigation__list-item {
       justify-content: flex-end;
 
-      /* Wide screens: show only first item */
-      @media (min-width: 1024px) {
-        &:not(:first-child, &--always-visible),
-        .indicator {
-          display: none;
+      body:has(main.container.container--fullwidth):has(
+          .denhaag-side-navigation
+        )
+        & {
+        @media (min-width: 1024px) {
+          &:not(&--always-visible),
+          .indicator {
+            display: none;
+          }
+        }
+      }
+
+      body:has(main.container:not(.container--fullwidth)):has(
+          .denhaag-side-navigation
+        )
+        & {
+        @media (min-width: 768px) {
+          &:not(&--always-visible),
+          .indicator {
+            display: none;
+          }
         }
       }
